Share Pins, Error, and re-exports
