+++ nrf52-hal-common/src/uarte/blocking.rs
@@ -13,16 +13,11 @@ use crate::target::{
 };
 
 use crate::prelude::*;
-use crate::gpio::{
-    p0::P0_Pin,
-    Output,
-    PushPull,
-};
-
-// Re-export SVD variants to allow user to directly set values
-pub use crate::target::uarte0::{
-    baudrate::BAUDRATEW as Baudrate,
-    config::PARITYW as Parity,
+use crate::uarte::{
+    Error,
+    Pins,
+    Parity,
+    Baudrate,
 };
 
 pub trait UarteExt: Deref<Target = uarte0::RegisterBlock> + Sized {
@@ -168,19 +163,3 @@ impl<T> Uarte<T> where T: UarteExt {
     }
 }
 
-
-pub struct Pins {
-    pub rxd: P0_Pin<Output<PushPull>>,
-    pub txd: P0_Pin<Output<PushPull>>,
-    pub cts: Option<P0_Pin<Output<PushPull>>>,
-    pub rts: Option<P0_Pin<Output<PushPull>>>,
-}
-
-
-#[derive(Debug)]
-pub enum Error {
-    TxBufferTooLong,
-    RxBufferTooLong,
-    Transmit,
-    Receive,
-}
