+++ nrf52-hal-common/src/uarte/nonblocking.rs
@@ -17,11 +17,6 @@ use crate::target::{
 };
 
 use crate::prelude::*;
-use crate::gpio::{
-    p0::P0_Pin,
-    Output,
-    PushPull,
-};
 
 use bbqueue::{
     Producer,
@@ -30,10 +25,11 @@ use bbqueue::{
     Error as BBQError
 };
 
-// Re-export SVD variants to allow user to directly set values
-pub use crate::target::uarte0::{
-    baudrate::BAUDRATEW as Baudrate,
-    config::PARITYW as Parity,
+use crate::uarte::{
+    Error,
+    Pins,
+    Parity,
+    Baudrate,
 };
 
 pub trait UarteAsyncExt: Deref<Target = uarte0::RegisterBlock> + Sized {
@@ -185,23 +181,6 @@ impl<T> UarteAsync<T> where T: UarteAsyncExt {
     pub fn free(self) -> T {
         self.periph
     }
-}
-
-
-pub struct Pins {
-    pub rxd: P0_Pin<Output<PushPull>>,
-    pub txd: P0_Pin<Output<PushPull>>,
-    pub cts: Option<P0_Pin<Output<PushPull>>>,
-    pub rts: Option<P0_Pin<Output<PushPull>>>,
-}
-
-
-#[derive(Debug)]
-pub enum Error {
-    TxBufferTooLong,
-    RxBufferTooLong,
-    Transmit,
-    Receive,
 }
 
 #[interrupt]
